outreach-tool-automation/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "outreach-tool-automation"
version = "0.1.0"
description = "Local-first multi-channel outreach automation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-dotenv>=1.0.1",
  "requests>=2.32.0",
  "pydantic>=2.7.0",
  "google-api-python-client>=2.133.0",
  "google-auth>=2.35.0",
  "gspread>=6.1.0",
  "firebase-admin>=6.5.0",
  "playwright>=1.45.0",
  "tenacity>=8.3.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "mypy>=1.10.0",
  "ruff>=0.5.0",
  "black>=24.4.0",
  "types-requests>=2.32.0.20240712",
  "pre-commit>=3.7.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]
extend-exclude = ["src/outreach_automation/templates"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "N", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["outreach_automation"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
mypy_path = "src"
exclude = "src/outreach_automation/templates/.*"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["tests"]
